+++ packages/orq-rc/src/orq_ai_sdk/contacts.py
@@ -87,6 +87,7 @@ class Contacts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                base_url=base_url or "",
                 operation_id="CreateContact",
                 oauth2_scopes=[],
                 security_source=get_security_from_env(
@@ -200,6 +201,7 @@ class Contacts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                base_url=base_url or "",
                 operation_id="CreateContact",
                 oauth2_scopes=[],
                 security_source=get_security_from_env(