+++ packages/orq-rc/src/orq_ai_sdk/feedback.py
@@ -78,6 +78,7 @@ class Feedback(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                base_url=base_url or "",
                 operation_id="CreateFeedback",
                 oauth2_scopes=[],
                 security_source=get_security_from_env(
@@ -182,6 +183,7 @@ class Feedback(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                base_url=base_url or "",
                 operation_id="CreateFeedback",
                 oauth2_scopes=[],
                 security_source=get_security_from_env(