+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -60,5 +60,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-11T15:58:00.983Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-02-13T15:49:16.783Z")
     r"""The date and time the resource was created"""